+++ example/main/example_cli.c
@@ -101,7 +101,7 @@ example_client_rpc(clicon_handle h,
     cva = cvec_find(cvv, "a"); /* get a cligen variable from vector */
     /* Create XML for example netconf RPC */
     if (clixon_xml_parse_va(YB_NONE, NULL, &xtop, NULL,
-			    "<rpc message-id=\"101\" xmlns=\"%s\" username=\"%s\">"
+			    "<rpc xmlns=\"%s\" username=\"%s\" message-id=\"101\">"
 			    "<example xmlns=\"urn:example:clixon\"><x>%s</x></example></rpc>",
 			    NETCONF_BASE_NAMESPACE,
 			    clicon_username_get(h),